+++ ldap/servers/slapd/back-ldbm/ldbm_modrdn.c
@@ -156,7 +156,8 @@ ldbm_back_modrdn( Slapi_PBlock *pb )
 
     if (pb->pb_conn)
     {
-        slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_modrdn", "enter conn=%" NSPRIu64 " op=%d\n", pb->pb_conn->c_connid, operation->o_opid);
+        slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_modrdn", "enter conn=%" NSPRIu64 " op=%d\n",
+                (long long unsigned int)pb->pb_conn->c_connid, operation->o_opid);
     }
 
     inst = (ldbm_instance *) be->be_instance_info;
@@ -1380,7 +1381,8 @@ common_return:
     }
     if (pb->pb_conn)
     {
-        slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_modrdn", "leave conn=%" NSPRIu64 " op=%d\n", pb->pb_conn->c_connid, operation->o_opid);
+        slapi_log_error (SLAPI_LOG_TRACE, "ldbm_back_modrdn", "leave conn=%" NSPRIu64 " op=%d\n",
+                (long long unsigned int)pb->pb_conn->c_connid, operation->o_opid);
     }
     return retval;
 }
